Add authorization on multiple endpoints

+++ src/main/java/web/javaproject/fooddeliveryapp/controller/RestaurantController.java
@@ -4,6 +4,9 @@ import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
@@ -13,11 +16,13 @@ import web.javaproject.fooddeliveryapp.dto.GetRestaurantDTO;
 import web.javaproject.fooddeliveryapp.dto.RestaurantDTO;
 import web.javaproject.fooddeliveryapp.mapper.RestaurantMapper;
 import web.javaproject.fooddeliveryapp.model.Restaurant;
+import web.javaproject.fooddeliveryapp.model.security.CustomUserDetails;
 import web.javaproject.fooddeliveryapp.service.RestaurantService;
 import web.javaproject.fooddeliveryapp.util.ValidationCheck;
 import org.springframework.ui.Model;
 
 import java.util.List;
+import java.util.Objects;
 
 @Controller
 @RequestMapping("/restaurant")
@@ -40,6 +45,19 @@ public class RestaurantController {
 
     @RequestMapping("/edit/{id}")
     public String edit(@PathVariable String id, Model model){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        GrantedAuthority firstAuthority = authentication.getAuthorities().stream().findFirst().orElse(null);
+        if(firstAuthority == null) {
+            return "access_denied";
+        }
+
+        if(Objects.equals(firstAuthority.getAuthority(), "RESTAURANT")) {
+            CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+
+            if(Long.parseLong(id) != userDetails.getAssociatedId()) {
+                return "acess_denied";
+            }
+        }
         model.addAttribute("restaurant", restaurantService.findById(Long.valueOf(id)));
         return "restaurantForm";
     }
@@ -49,6 +67,20 @@ public class RestaurantController {
                                BindingResult bindingResult,
                                Model model
     ){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        GrantedAuthority firstAuthority = authentication.getAuthorities().stream().findFirst().orElse(null);
+        if(firstAuthority == null) {
+            return "access_denied";
+        }
+
+        if(Objects.equals(firstAuthority.getAuthority(), "RESTAURANT")) {
+            CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+
+            if(restaurant.getId() != userDetails.getAssociatedId()) {
+                return "acess_denied";
+            }
+        }
+
         if (bindingResult.hasErrors()){
             model.addAttribute("restaurant", restaurant);
             return "restaurantForm";
@@ -56,12 +88,25 @@ public class RestaurantController {
 
         restaurantService.save(restaurant);
 
-
-        return "redirect:/restaurant" ;
+        return "redirect:/restaurant";
     }
 
     @RequestMapping("/delete/{id}")
     public String deleteById(@PathVariable String id){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        GrantedAuthority firstAuthority = authentication.getAuthorities().stream().findFirst().orElse(null);
+        if(firstAuthority == null) {
+            return "access_denied";
+        }
+
+        if(Objects.equals(firstAuthority.getAuthority(), "RESTAURANT")) {
+            CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+
+            if(Long.parseLong(id) != userDetails.getAssociatedId()) {
+                return "acess_denied";
+            }
+        }
+
         restaurantService.deleteById(Long.valueOf(id));
         return "redirect:/restaurant";
     }
@@ -74,39 +119,4 @@ public class RestaurantController {
         model.addAttribute("restaurantsAll", restaurantsAll);
         return "restaurantForm";
     }
-
-
-
-
-
-
-//    @GetMapping("/get")
-//    public ResponseEntity<?> getByEmail(@RequestParam String restaurantEmail) {
-//        try {
-//            Restaurant restaurant = restaurantService.getRestaurantByEmail(restaurantEmail);
-//            GetRestaurantDTO restaurantDTO = restaurantMapper.toGetDto(restaurant);
-//
-//            return new ResponseEntity<>(restaurantDTO, HttpStatus.OK);
-//        } catch (Exception e) {
-//            return new ResponseEntity<>("Error retrieving restaurant: " + e.getMessage(), HttpStatus.BAD_REQUEST);
-//        }
-//    }
-//
-//    @PostMapping("/create")
-//    public ResponseEntity<?> create(@RequestBody @Valid CreateRestaurantDTO createRestaurantDTO, BindingResult bindingResult) {
-//        if (bindingResult.hasErrors()) {
-//            String errorMessage = ValidationCheck.extractValidationErrorMessage(bindingResult);
-//            return ResponseEntity.badRequest().body(errorMessage);
-//        }
-//
-//        try {
-//            Restaurant restaurant = restaurantMapper.createDTOtoEntity(createRestaurantDTO);
-//            Restaurant createdRestaurant = restaurantService.createRestaurant(restaurant);
-//            RestaurantDTO createdRestaurantDTO = restaurantMapper.toCreateDto(createdRestaurant);
-//
-//            return new ResponseEntity<>(createdRestaurantDTO, HttpStatus.CREATED);
-//        } catch (Exception e) {
-//            return new ResponseEntity<>("Error creating client: " + e.getMessage(), HttpStatus.BAD_REQUEST);
-//        }
-//    }
 }
